+++ analyze_mft/constants/constants.py
